Started HLG for array slicing
